refactor: update `SpeakersSection` and abstract `SpeakersContent`
- Extract out SpeakersContent so that SpeakersProvider is able to set up 
context
